requirements.txt
pandas~=2.1.1
numpy~=1.25.2
pydicom~=2.4.3
xpinyin~=0.7.6
torchvision~=0.15.2
pillow~=9.4.0
easydict~=1.10
yaml~=0.2.5
pyyaml~=6.0.1
tqdm~=4.66.1
matplotlib~=3.8.0
scikit-learn~=1.3.1
accelerate~=0.23.0
opencv-python~=4.8.0.76
monai~=1.2.0
scipy~=1.11.2
datasets~=2.14.5
evaluate~=0.4.0
seaborn~=0.13.2
timm~=0.9.7
pytz~=2023.3.post1
mmeval~=0.2.1
objprint~=0.2.2
torchinfo~=1.8.0